requirements.txt
# latest revision
celery
Django
django-celery
django-nose
django-queued-storage
fluent-logger
mock
MySQL-python
pycrypto
python-memcached
south
tastypie-rpc-proxy
tox

# specific revision
-e git+https://github.com/toastdriven/django-tastypie.git#egg=django-tastypie
